+++ src/rics/translation/_from_config.py
@@ -1,81 +1,118 @@
-from typing import TYPE_CHECKING, Any, Dict, Union
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Optional, Tuple, Type, TypeVar, Union
 
-import yaml
+import toml
 
-from rics._internal_support.types import PathLikeType
 from rics.mapping import Mapper
+from rics.translation import fetching
 from rics.translation.exceptions import ConfigurationError
-from rics.translation.fetching import Fetcher, PandasFetcher, SqlFetcher
-from rics.translation.offline import PlaceholderOverrides
-from rics.utility.misc import tname
+from rics.translation.offline import DefaultTranslations, PlaceholderOverrides
 
 if TYPE_CHECKING:
     from rics.translation._translator import Translator  # pragma: no cover
 
+FetcherFactory = Callable[[str, Dict[str, Any]], fetching.Fetcher]
 
-def translator_from_yaml_config(path: Union[PathLikeType, Dict[str, Any]]) -> "Translator":
-    """Create a translator from a YAML file.
-
-    Args:
-        path: Path to a YAML file, or a pre-parsed dict.
-
-    Returns:
-        A Translator object.
-
-    Raises:
-        ConfigurationError: If the config is invalid.
-    """
-    if isinstance(path, dict):
-        config = path  # pragma: no cover
-    else:
-        with open(path) as f:
-            config = yaml.safe_load(f)
-
-    try:
-        return _parse(config)
-    except Exception as e:  # noqa: B902, pragma: no cover
-        raise ConfigurationError(f"{e}, {path=}") from e
+T = TypeVar("T", DefaultTranslations, PlaceholderOverrides)
 
 
-def _parse(config: Dict[str, Any]) -> "Translator":
-    main = config.pop("translator")
-    fetcher = config.pop("fetcher")
-    mapper = _make_mapper(**config.pop("name-to-source-mapping", {}))
-    placeholder_overrides = PlaceholderOverrides.from_dict(config.pop("placeholder-overrides", {}))
-    default_translations = config.pop("default-translations", None)
+def _default_factory(name: str, kwargs: Dict[str, Any]) -> fetching.Fetcher:
+    fetcher_clazz = getattr(fetching, name, None)
+    if fetcher_clazz is None:
+        raise ValueError(f"Fetcher class '{name}' not known. Consider using the 'fetcher_factory' argument.")
+    fetcher = fetcher_clazz(**kwargs)
+    if not isinstance(fetcher, fetching.Fetcher):
+        raise TypeError(fetcher)
+    return fetcher
 
-    if config:
-        raise ConfigurationError(f"Invalid configuration. Unknown keys: {list(config)}")
 
-    fetcher = _make_fetcher(fetcher, placeholder_overrides)
+def translator_from_toml_config(f: str, fetcher_factory: FetcherFactory = None) -> "Translator":
+    """Create a Translator from TOML specification."""
+    dict_config: Dict[str, Any] = toml.load(f)
 
-    from rics.translation._translator import Translator
+    _check_allowed_keys(["translator", "mapping", "fetching", "unknown_ids"], dict_config, "<root>")
 
-    return Translator(fetcher, **main, mapper=mapper, default_translations=default_translations)
+    translator_config = dict_config.pop("translator", {})
+    fetcher = _make_fetcher(fetcher_factory or _default_factory, **dict_config.pop("fetching"))
+    mapper = _make_mapper(**translator_config.pop("mapping", {}))
+    default_fmt, default_translations = _make_default_translations(**dict_config.pop("unknown_ids", {}))
+
+    from rics.translation import Translator
+
+    return Translator(
+        fetcher, mapper=mapper, default_translations=default_translations, default_fmt=default_fmt, **translator_config
+    )
+
+
+def _make_default_translations(**config: Any) -> Tuple[str, Optional[DefaultTranslations]]:
+    _check_allowed_keys(["fmt", "overrides"], config, toml_path="translator.unknown_ids")
+    return (
+        config.pop("fmt", None),
+        _create_overrides(config["overrides"], DefaultTranslations) if "overrides" in config else None,
+    )
 
 
 def _make_mapper(**config: Any) -> Mapper:
-    if "score-function" in config:  # pragma: no cover
-        score_function = config.pop("score-function")
-        config["score_function"] = score_function.pop("name")
-        config.update(**score_function)
+    _check_forbidden_keys(["candidates", "cardinality"], config, "fetcher.mapping")
 
-    return Mapper.from_dict(config)
+    score_function = config.pop("score_function", {})
+    if len(score_function) > 1:
+        raise ConfigurationError(f"At most one score function may be specified, but got: {sorted(score_function)}")
+    score_function_name, score_function_kwargs = next(iter(score_function.items()))
+
+    return Mapper(score_function=score_function_name, **score_function_kwargs, **config)
 
 
-def _make_fetcher(base_config: Dict[str, Any], placeholder_overrides: PlaceholderOverrides) -> "Fetcher":
-    name = base_config.pop("class")
-
-    clazz = None
-    if "." not in name:
-        for candidate in (PandasFetcher, SqlFetcher):
-            if tname(candidate) == name:
-                clazz = candidate
+def _make_fetcher(factory: FetcherFactory, **config: Any) -> fetching.Fetcher:
+    overrides: Optional[PlaceholderOverrides]
+    if "mapping" in config:
+        mapping = config.pop("mapping")
+        _check_allowed_keys("overrides", mapping, toml_path="fetching.mapping")
+        overrides = _create_overrides(mapping.pop("overrides", {}), PlaceholderOverrides)
     else:
-        raise NotImplementedError("TODO: Custom fetcher class loading with YAML.")
+        overrides = None
 
-    if clazz is None:
-        raise ValueError(f"Could not find a fetcher class with {name=}.")  # pragma: no cover
+    if len(config) == 0:
+        raise ConfigurationError("Fetcher implementation section missing.")
+    if len(config) > 1:
+        raise ConfigurationError(f"Multiple fetcher implementations specified: {sorted(config)}.")
 
-    class_specific_kwargs = base_config.pop("class-specific-kwargs", {})
-    return clazz(**base_config, **class_specific_kwargs, placeholder_overrides=placeholder_overrides)
+    clazz, kwargs = next(iter(config.items()))
+    _check_forbidden_keys(
+        ["mapping", "overrides", "placeholder_overrides"], kwargs, toml_path=f"fetching.{clazz}"
+    )  # No algorithmic matching yet
+    kwargs["placeholder_overrides"] = overrides
+    return factory(clazz, kwargs)
+
+
+def _create_overrides(dict_config: Dict[str, Any], clazz: Type[T]) -> T:
+    shared: Dict[str, Any] = {}
+    source_specific: Dict[str, Dict[str, Any]] = {}
+
+    for outer_key, outer_value in dict_config.items():
+        if isinstance(outer_value, dict):
+            for k, v in outer_value.items():
+                if outer_key not in source_specific:
+                    source_specific[outer_key] = {}
+                source_specific[outer_key][k] = v
+        else:
+            shared[outer_key] = outer_value
+
+    return clazz(shared=shared, source_specific=source_specific)
+
+
+def _check_forbidden_keys(forbidden: Union[str, Iterable[str]], actual: Iterable[str], toml_path: str) -> None:
+    if isinstance(forbidden, str):
+        forbidden = [forbidden]
+
+    bad_keys = set(forbidden).intersection(actual)
+    if bad_keys:
+        raise ValueError(f"Forbidden keys {sorted(bad_keys)} in [{toml_path}]-section.")
+
+
+def _check_allowed_keys(allowed: Iterable[str], actual: Iterable[str], toml_path: str) -> None:
+    if isinstance(allowed, str):
+        allowed = [allowed]
+
+    bad_keys = set(actual).difference(allowed)
+    if bad_keys:
+        raise ValueError(f"Forbidden keys {sorted(bad_keys)} in [{toml_path}]-section.")